feat(cli): 添加命令行参数解析与安装目标
【更新内容】
- 引入 getopt_long 实现 POSIX 风格命令行解析
- 支持 -h/--help 显示帮助信息
- 支持 -v/--version 显示版本信息
- 解析两个位置参数：源文件路径、目标路径
- 将目标文件名从 bin/copy 改为 bin/ufcp
- 添加 CXXFLAGS 启用 C++17 标准与 O2 优化
- 添加 install/uninstall 安装目标
- 支持 PREFIX 自定义安装路径
- 支持 DESTDIR 适配打包工具

【技术决策】(如无重大决策可留空)
- 问题背景：当前 main 函数使用硬编码路径，无法通用；Makefile 仅有编译功能
- 决策依据：采用 POSIX getopt_long，兼容性好；采用 GNU Coding Standards 的 PREFIX 约定

【优缺点分析】(如无重构可留空)
- 优点：标准化命令行接口，用户可通过 make install 直接安装
- 缺点：无

+++ src/UltraFastCopy.cpp
@@ -50,7 +50,7 @@ uint64_t UltraFastCopy::GetFileSize(const std::string& filePath) {
 	if (file.fail()) throw std::runtime_error("定位文件末尾失败: " + filePath);
 	std::streampos pos = file.tellg();
 	if (pos == std::streampos(-1)) throw std::runtime_error("获取文件大小失败: " + filePath);
-	if (pos > std::numeric_limits<uint64_t>::max()) throw std::overflow_error("文件大小≥ 2^34 GiB" + filePath);
+	if (static_cast<uint64_t>(pos) > std::numeric_limits<uint64_t>::max()) throw std::overflow_error("文件大小≥ 2^34 GiB" + filePath);
 	return static_cast<uint64_t>(pos);
 }
 //计算当前系统最优读写大小，即单次read/write的大小
